pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "banklab"
version = "0.1.0"
description = "JPM vs Morgan Stanley analytics platform using public data"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Leo Tabatabaei", email = "leo@example.com" }
]
keywords = ["finance", "banking", "sec", "edgar", "quantitative-finance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "pandas>=2.0",
    "pyarrow>=14.0",
    "requests>=2.31",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1.6",
    "ipykernel>=6.26",
]
analysis = [
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "statsmodels>=0.14",
    "scipy>=1.11",
]

[project.scripts]
banklab = "banklab.run:main"

[project.urls]
Homepage = "https://github.com/Leotaby/banklab"
Documentation = "https://github.com/Leotaby/banklab#readme"
Repository = "https://github.com/Leotaby/banklab"

[tool.hatch.build.targets.wheel]
packages = ["src/banklab"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["banklab"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]
